+++ src/types/rpc.ts
@@ -1,5 +1,6 @@
 import type { Address } from 'abitype'
 
+import type { RpcAuthorizationList } from '../experimental/eip7702/types/rpc.js'
 import type {
   Block,
   BlockIdentifier,
@@ -15,11 +16,13 @@ import type {
   TransactionEIP1559,
   TransactionEIP2930,
   TransactionEIP4844,
+  TransactionEIP7702,
   TransactionLegacy,
   TransactionReceipt,
   TransactionRequestEIP1559,
   TransactionRequestEIP2930,
   TransactionRequestEIP4844,
+  TransactionRequestEIP7702,
   TransactionRequestLegacy,
 } from './transaction.js'
 import type { Omit, OneOf, PartialBy } from './utils.js'
@@ -27,7 +30,13 @@ import type { Omit, OneOf, PartialBy } from './utils.js'
 export type Index = `0x${string}`
 export type Quantity = `0x${string}`
 export type Status = '0x0' | '0x1'
-export type TransactionType = '0x0' | '0x1' | '0x2' | (string & {})
+export type TransactionType =
+  | '0x0'
+  | '0x1'
+  | '0x2'
+  | '0x3'
+  | '0x4'
+  | (string & {})
 
 export type RpcBlock<
   blockTag extends BlockTag = BlockTag,
@@ -52,6 +61,10 @@ export type RpcTransactionRequest = OneOf<
   | TransactionRequestEIP2930<Quantity, Index, '0x1'>
   | TransactionRequestEIP1559<Quantity, Index, '0x2'>
   | TransactionRequestEIP4844<Quantity, Index, '0x3'>
+  | (Omit<
+      TransactionRequestEIP7702<Quantity, Index, '0x4'>,
+      'authorizationList'
+    > & { authorizationList?: RpcAuthorizationList | undefined })
 >
 // `yParity` is optional on the RPC type as some nodes do not return it
 // for 1559 & 2930 transactions (they should!).
@@ -67,6 +80,13 @@ export type RpcTransaction<pending extends boolean = boolean> = OneOf<
     >
   | PartialBy<
       Omit<TransactionEIP4844<Quantity, Index, pending, '0x3'>, 'typeHex'>,
+      'yParity'
+    >
+  | PartialBy<
+      Omit<
+        TransactionEIP7702<Quantity, Index, pending, '0x4'>,
+        'authorizationList' | 'typeHex'
+      > & { authorizationList?: RpcAuthorizationList | undefined },
       'yParity'
     >
 >