+++ src/utils/transaction/parseTransaction.test.ts
@@ -7,6 +7,7 @@ import type {
   TransactionSerializableEIP1559,
   TransactionSerializableEIP2930,
   TransactionSerializableEIP4844,
+  TransactionSerializableEIP7702,
   TransactionSerializableLegacy,
 } from '../../types/transaction.js'
 import { toHex } from '../encoding/toHex.js'
@@ -15,6 +16,8 @@ import { keccak256 } from '../hash/keccak256.js'
 import { parseEther } from '../unit/parseEther.js'
 import { parseGwei } from '../unit/parseGwei.js'
 
+import type { Address } from 'abitype'
+import { wagmiContractConfig } from '../../../test/src/abis.js'
 import {
   parseAccessList,
   parseTransaction,
@@ -27,6 +30,220 @@ const base = {
   nonce: 785,
   value: parseEther('1'),
 } satisfies TransactionSerializableBase
+
+describe('eip7702', () => {
+  const baseEip7702 = {
+    ...base,
+    authorizationList: [
+      {
+        contractAddress: wagmiContractConfig.address.toLowerCase() as Address,
+        chainId: 1,
+        nonce: 420,
+        r: '0x60fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fe',
+        s: '0x60fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fe',
+        v: 27n,
+        yParity: 0,
+      },
+      {
+        contractAddress: '0x0000000000000000000000000000000000000000',
+        chainId: 10,
+        nonce: 69,
+        r: '0x60fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fe',
+        s: '0x60fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fe',
+        v: 28n,
+        yParity: 1,
+      },
+    ],
+    chainId: 1,
+  } as const satisfies TransactionSerializableEIP7702
+
+  test('default', () => {
+    const serialized = serializeTransaction(baseEip7702)
+    const transaction = parseTransaction(serialized)
+    assertType<TransactionSerializableEIP7702>(transaction)
+    expect(transaction).toEqual({ ...baseEip7702, type: 'eip7702' })
+  })
+
+  test('args: fees', () => {
+    const args = {
+      ...baseEip7702,
+      maxFeePerGas: parseGwei('2'),
+      maxPriorityFeePerGas: parseGwei('1'),
+    }
+    const serialized = serializeTransaction(args)
+    const transaction = parseTransaction(serialized)
+    assertType<TransactionSerializableEIP7702>(transaction)
+    expect(transaction).toEqual({ ...args, type: 'eip7702' })
+  })
+
+  test('args: gas', () => {
+    const args = {
+      ...baseEip7702,
+      gas: 69n,
+    }
+    const serialized = serializeTransaction(args)
+    const transaction = parseTransaction(serialized)
+    assertType<TransactionSerializableEIP7702>(transaction)
+    expect(transaction).toEqual({ ...args, type: 'eip7702' })
+  })
+
+  test('args: accessList', () => {
+    const args = {
+      ...baseEip7702,
+      accessList: [
+        {
+          address: '0x0000000000000000000000000000000000000000',
+          storageKeys: [
+            '0x0000000000000000000000000000000000000000000000000000000000000001',
+            '0x60fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fe',
+          ],
+        },
+      ],
+    } as const satisfies TransactionSerializableEIP7702
+    const serialized = serializeTransaction(args)
+    const transaction = parseTransaction(serialized)
+    assertType<TransactionSerializableEIP7702>(transaction)
+    expect(transaction).toEqual({ ...args, type: 'eip7702' })
+  })
+
+  test('args: data', () => {
+    const args = {
+      ...baseEip7702,
+      data: '0x1234',
+    } satisfies TransactionSerializableEIP7702
+    const serialized = serializeTransaction(args)
+    const transaction = parseTransaction(serialized)
+    assertType<TransactionSerializableEIP7702>(transaction)
+    expect(transaction).toEqual({ ...args, type: 'eip7702' })
+  })
+
+  test('signed', async () => {
+    const signature = await sign({
+      hash: keccak256(serializeTransaction(baseEip7702)),
+      privateKey: accounts[0].privateKey,
+    })
+    const serialized = serializeTransaction(baseEip7702, signature)
+    expect(parseTransaction(serialized)).toEqual({
+      ...baseEip7702,
+      ...signature,
+      type: 'eip7702',
+      yParity: 0,
+    })
+  })
+
+  describe('errors', () => {
+    test('invalid access list (invalid address)', () => {
+      expect(() =>
+        parseTransaction(
+          `0x04${toRlp([
+            toHex(1), // chainId
+            toHex(0), // nonce
+            toHex(1), // maxPriorityFeePerGas
+            toHex(1), // maxFeePerGas
+            toHex(1), // gas
+            '0x0000000000000000000000000000000000000000', // to
+            toHex(0), // value
+            '0x', // data
+            [
+              [
+                '0x',
+                [
+                  '0x0000000000000000000000000000000000000000000000000000000000000001',
+                ],
+              ],
+            ], // accessList
+            [], // authorizationList
+          ]).slice(2)}`,
+        ),
+      ).toThrowErrorMatchingInlineSnapshot(`
+        [InvalidAddressError: Address "0x" is invalid.
+
+        - Address must be a hex value of 20 bytes (40 hex characters).
+        - Address must match its checksum counterpart.
+
+        Version: viem@x.y.z]
+      `)
+
+      expect(() =>
+        parseTransaction(
+          `0x04${toRlp([
+            toHex(1), // chainId
+            toHex(0), // nonce
+            toHex(1), // maxPriorityFeePerGas
+            toHex(1), // maxFeePerGas
+            toHex(1), // gas
+            '0x0000000000000000000000000000000000000000', // to
+            toHex(0), // value
+            '0x', // data
+            [['0x123456', ['0x0']]], // accessList
+            [], // authorizationList
+          ]).slice(2)}`,
+        ),
+      ).toThrowErrorMatchingInlineSnapshot(`
+        [InvalidAddressError: Address "0x123456" is invalid.
+
+        - Address must be a hex value of 20 bytes (40 hex characters).
+        - Address must match its checksum counterpart.
+
+        Version: viem@x.y.z]
+      `)
+    })
+
+    test('invalid transaction (all missing)', () => {
+      expect(() =>
+        parseTransaction(`0x04${toRlp([]).slice(2)}`),
+      ).toThrowErrorMatchingInlineSnapshot(`
+        [InvalidSerializedTransactionError: Invalid serialized transaction of type "eip7702" was provided.
+
+        Serialized Transaction: "0x04c0"
+        Missing Attributes: chainId, nonce, maxPriorityFeePerGas, maxFeePerGas, gas, to, value, data, accessList, authorizationList
+
+        Version: viem@x.y.z]
+      `)
+    })
+
+    test('invalid transaction (some missing)', () => {
+      expect(() =>
+        parseTransaction(`0x04${toRlp(['0x0', '0x1']).slice(2)}`),
+      ).toThrowErrorMatchingInlineSnapshot(`
+        [InvalidSerializedTransactionError: Invalid serialized transaction of type "eip7702" was provided.
+
+        Serialized Transaction: "0x04c20001"
+        Missing Attributes: maxPriorityFeePerGas, maxFeePerGas, gas, to, value, data, accessList, authorizationList
+
+        Version: viem@x.y.z]
+      `)
+    })
+
+    test('invalid transaction (missing signature)', () => {
+      expect(() =>
+        parseTransaction(
+          `0x04${toRlp([
+            '0x',
+            '0x',
+            '0x',
+            '0x',
+            '0x',
+            '0x',
+            '0x',
+            '0x',
+            '0x',
+            '0x',
+            '0x',
+            '0x',
+          ]).slice(2)}`,
+        ),
+      ).toThrowErrorMatchingInlineSnapshot(`
+        [InvalidSerializedTransactionError: Invalid serialized transaction of type "eip7702" was provided.
+
+        Serialized Transaction: "0x04cc808080808080808080808080"
+        Missing Attributes: s
+
+        Version: viem@x.y.z]
+      `)
+    })
+  })
+})
 
 describe('eip4844', () => {
   const baseEip4844 = {
@@ -80,7 +297,7 @@ describe('eip4844', () => {
           ],
         },
       ],
-    } satisfies TransactionSerializableEIP4844
+    } as const satisfies TransactionSerializableEIP4844
     const serialized = serializeTransaction(args)
     const transaction = parseTransaction(serialized)
     assertType<TransactionSerializableEIP4844>(transaction)