+++ test/setup.ts
@@ -5,17 +5,25 @@ import { cleanupCache, listenersCache } from '~viem/utils/observe.js'
 import { promiseCache, responseCache } from '~viem/utils/promise/withCache.js'
 import { socketClientCache } from '~viem/utils/rpc/socket.js'
 
+import { setErrorConfig } from '../src/errors/base.js'
 import * as instances from './src/anvil.js'
 
 const client = instances.anvilMainnet.getClient()
 
 beforeAll(() => {
+  setErrorConfig({
+    getDocsUrl({ docsBaseUrl, docsPath }) {
+      return docsPath
+        ? `${docsBaseUrl ?? 'https://viem.sh'}${docsPath}`
+        : undefined
+    },
+    version: 'viem@x.y.z',
+  })
   vi.mock('../src/errors/utils.ts', () => ({
     getContractAddress: vi
       .fn()
       .mockReturnValue('0x0000000000000000000000000000000000000000'),
     getUrl: vi.fn().mockReturnValue('http://localhost'),
-    getVersion: vi.fn().mockReturnValue('viem@x.y.z'),
   }))
 })