+++ test/src/anvil.ts
@@ -5,7 +5,7 @@ import {
   optimism,
   optimismSepolia,
   sepolia,
-  zkSync,
+  zksync,
 } from '../../src/chains/index.js'
 import { ipc } from '../../src/clients/transports/ipc.js'
 import {
@@ -61,8 +61,8 @@ export const anvilOptimismSepolia = defineAnvil({
   port: 8945,
 })
 
-export const anvilZkSync = defineAnvil({
-  chain: zkSync,
+export const anvilZksync = defineAnvil({
+  chain: zksync,
   forkUrl: getEnv(
     'VITE_ANVIL_FORK_URL_ZKSYNC',
     'https://mainnet.era.zksync.io',
@@ -248,6 +248,7 @@ function defineAnvil<const chain extends Chain>(
         instance: anvil({
           forkUrl,
           forkBlockNumber,
+          hardfork: 'Prague',
           ...options,
         }),
         port,